+++ addon/components/paper-sidenav-container.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+const { Component } = Ember;
+
+export default Component.extend({
+  classNames: ['flex', 'layout-row'],
+  attributeBindings: ['style'],
+  style: Ember.String.htmlSafe('overflow: hidden; position: relative')
+});