+++ app/services/constants.js
@@ -44,19 +44,28 @@ export default Ember.Service.extend({
   }),
 
   MEDIA: {
-    'sm': '(max-width: 599px)',
-    'gt-sm': '(min-width: 600px)',
-    'md': '(min-width: 600px) and (max-width: 959px)',
-    'gt-md': '(min-width: 960px)',
-    'lg': '(min-width: 960px) and (max-width: 1199px)',
-    'gt-lg': '(min-width: 1200px)'
+    'xs': '(max-width: 599px)',
+    'gt-xs': '(min-width: 600px)',
+    'sm': '(min-width: 600px) and (max-width: 959px)',
+    'gt-sm': '(min-width: 960px)',
+    'md': '(min-width: 960px) and (max-width: 1279px)',
+    'gt-md': '(min-width: 1280px)',
+    'lg': '(min-width: 1280px) and (max-width: 1919px)',
+    'gt-lg': '(min-width: 1920px)',
+    'xl': '(min-width: 1920px)',
+    'print': 'print'
   },
+  
   MEDIA_PRIORITY: [
+    'xl',
     'gt-lg',
     'lg',
     'gt-md',
     'md',
     'gt-sm',
-    'sm'
+    'sm',
+    'gt-xs',
+    'xs',
+    'print'
   ]
 });