+++ tests/dummy/app/controllers/demo/sidenav.js
@@ -1,0 +1,10 @@
+import Ember from 'ember';
+const { Controller } = Ember;
+
+export default Controller.extend({
+  actions: {
+    toggleSourceCode() {
+      this.toggleProperty('showSourceCode');
+    }
+  }
+});